+++ server/server.js
@@ -4,6 +4,7 @@ const cors = require('cors');
 const sequelize = require('./config/database');
 const authRoutes = require('./routes/authRoutes');
 const discRoutes = require('./routes/discRoutes');
+const masterRoutes = require('./routes/masterRoutes');
 const app = express();
 
 app.use(cors());
@@ -11,6 +12,7 @@ app.use(express.json());
 
 app.use('/', discRoutes);
 app.use('/auth', authRoutes);
+app.use('/master', masterRoutes);
 
 sequelize.sync()
     .then(() => {